04-monitoring/app/requirements.txt
streamlit
elasticsearch
openai
psycopg2-binary
python-dotenv